Replaced Long with Integer for optimization

+++ Java/miscellaneous/HashingPrefixPostfix.java
@@ -8,9 +8,9 @@ public class HashingPrefixPostfix {
     //Leet code problem https://leetcode.com/problems/subarray-sum-equals-k/
     public static int subarraySum(int[] nums, int k) {
         int result = 0;
-        long currentSum = 0;
-        Map<Long,Integer> prefixSum = new HashMap<>();
-        prefixSum.put(0L,1);
+        int currentSum = 0;
+        Map<Integer,Integer> prefixSum = new HashMap<>();
+        prefixSum.put(0,1);
 
         for(int i = 0;i<nums.length;i++)
         {
